Refactor script to follow OOP princples
Encapsulate functionality related to interacting with the Talent
REST API within the `TalentSolarMonitor` class.

The script still accepts command-line arguments for username,
password, and JSON output, but now uses an instance of the
`TalentSolarMonitor` class to handle the API interactions and
data retrieval.

Fixes Convert script to use OOP paradigms #1

Signed-off-by: Lenz Grimmer <lenz@grimmer.com>